pyproject.toml
[tool.poetry]
name = "flask_assets_tutorial"
version = "0.1.1"
description = "Flask application which compiles ands serves frontend assets with Flask-Assets."
authors = ["Todd Birchard <toddbirchard@gmail.com>"]
maintainers = ["Todd Birchard <toddbirchard@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://hackersandslackers.com/flask-assets/"
repository = "https://github.com/hackersandslackers/flask-assets-tutorial/"
documentation = "https://hackersandslackers.com/flask-assets/"
keywords = ["Flask", "Flask-Assets", "Bundles", "Frontend", "Tutorial"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
flask = "*"
flask-assets = "*"
lesscpy = "*"
cssmin = "*"
jsmin = "*"
python-dotenv = "*"
pytest = "*"
black = "*"
isort = "*"
gunicorn = "*"
flask-sock = "^0.7.0"

[tool.poetry.scripts]
run = "wsgi:app"

[tool.poetry.urls]
issues = "https://github.com/hackersandslackers/flask-assets-tutorial/issues"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
src_paths = ["flask_assets_tutorial", "config", "wsgi"]

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["flask_login_tutorial", "config", "wsgi"]

[tool.pylint.'MESSAGES CONTROL']
disable = "C0103,C0301,W0703,W0621"
